% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MSIS.R
\name{calculate_MSIS}
\alias{calculate_MSIS}
\title{Benchmarking the perforamance of Similarity against ETS and SHD}
\usage{
calculate_MSIS(tsi, gamma = 0.05, fh, nts = 100, Dist_type = 3)
}
\arguments{
\item{tsi}{ith particular reference series}

\item{gamma}{calibrated significance level quantiles}

\item{fh}{number of forecasting horizon}

\item{nts}{number of the most similar series}

\item{Dist_type}{number of forecasting benchmarks}
}
\value{
Dataframe containing theaccuracy of MSIS, Coverage, Upper coverage, Spread and MASE
using ETS, SHD, Similarity and ETS-Similarity separately
}
\description{
Benchmarking the perforamance of Similarity against ETS and SHD
}
\references{
Svetunkov & Petropoulos (2018)
}
\author{

}
