% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MSIS.R
\name{calculate_PI_MSIS}
\alias{calculate_PI_MSIS}
\title{Calculate the mean scaled interval score(MSIS) to evaluate the performance of generated predictive intervals}
\usage{
calculate_PI_MSIS(
  PIL,
  PIU,
  y,
  yout,
  gamma = 0.05,
  scaling = c("sdiff", "snaive")
)
}
\arguments{
\item{PIL}{lower bounds for the prediction interval}

\item{PIU}{upper bounds for the prediction interval}

\item{y}{"vector" the observation series}

\item{yout}{"vector" the generated series}

\item{gamma}{calibrated significance level quantiles}

\item{scaling}{scaling method: using different scaling or not}
}
\value{
The value calculated by MSIS function
}
\description{
Calculate the mean scaled interval score(MSIS) to evaluate the performance of generated predictive intervals
}
\author{

}
