% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropoutMechanism.R
\name{CreateNewDropoutMechanism}
\alias{CreateNewDropoutMechanism}
\title{A function which creates a DropOut Mechanism object}
\usage{
CreateNewDropoutMechanism(type, text, cols.needed = vector("character"),
  GetDropTime, parameters = NULL)
}
\arguments{
\item{type}{The type of mechanism (e.g. "MCAR" or "MNAR")}

\item{text}{A short string describing the mechanism (only used for printing)}

\item{cols.needed}{Which columns in the SingleSim$data data frame must be included for this
drop out mechanism to work. This option could allow drop out mechanism which depend on covariates to be
included.}

\item{GetDropTime}{A function with two arguments event.times and data, the corresponding entries from the 
SingleSim object. This function should return a list of dropout times (if a subject does not dropout its dropout time should be 
their current censored.time (i.e. the study follow up time))}

\item{parameters}{A list of named parameters for the mechanism (only used for printing) or NULL if none}
}
\description{
A function which creates a DropOut Mechanism object
}
\seealso{
\code{\link{DropoutMechanism.object}}
}

