% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scenario.R
\name{CreateScenario}
\alias{CreateScenario}
\title{Boo}
\usage{
CreateScenario(object, description = "")
}
\arguments{
\item{object}{TODO}

\item{description}{TODO}
}
\value{
TODO
}
\description{
Boo
}

