% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expander.R
\name{expandEventCount}
\alias{expandEventCount}
\title{Expand event counts into a list of event times}
\usage{
expandEventCount(count, time)
}
\arguments{
\item{count}{a vector of event counts.  All entries must be
non-negative.}

\item{time}{a matching (strictly positive) vector of followup
times.}
}
\value{
a list of vectors of event times
}
\description{
This function exists to allow clinical trial data which typically
gives event counts over time to be plugged into this software, which
relies on actual event counts.
}
\details{
This function always produces a warning: anyone relying on this
function to actually analyze data should take great care.
}
\examples{
expandEventCount(count=c(0, 20), time=c(10, 20))
}
