% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{extract_results}
\alias{extract_results}
\title{Extract the results of running a scenario}
\usage{
extract_results(answer, name, description)
}
\arguments{
\item{answer}{A named list of lists}

\item{name}{The name of the lists of answer which should be extracted and 
put together into a \code{sc}}

\item{description}{The description parameter to be passed into the \code{CreateScenario} function}
}
\value{
A \code{Scenario} object
}
\description{
This function is a wrapper around \code{\link{CreateScenario}}
See the user guide vignette for an example of using this function
}
\seealso{
\code{\link{CreateScenario}}
}
