% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ImputeSim.R
\name{GetImputedDataSet}
\alias{GetImputedDataSet}
\title{TODO}
\usage{
GetImputedDataSet(imputeSim, index)
}
\arguments{
\item{imputeSim}{TODO}

\item{index}{TODO}
}
\value{
TODO
}
\description{
TODO
}

