% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit.R
\name{Impute}
\alias{Impute}
\title{Produce imputed data sets}
\usage{
Impute(fit, impute.mechanism, N)
}
\arguments{
\item{fit}{A \code{SingleSimFit} object}

\item{impute.mechanism}{An \code{ImputeMechanism} object}

\item{N}{The number of data sets to impute}
}
\value{
An \code{ImputeSim} object
}
\description{
Given a \code{SingleSimFit} object and an imputation mechanism,
create a collection of imputed data sets
}

