% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/specificImputeMechanisms.R
\name{j2r}
\alias{j2r}
\title{Create a j2r \code{ImputeMechanism} object}
\usage{
j2r()
}
\value{
An \code{ImputeMechanism} object
}
\description{
Imputation using this mechanism will follow the jump to reference
(j2r) model whereby missing counts for subjects in both arms will
be imputed according to the mean of the placebo arm conditioned
on the subject's observed number of events
}
\seealso{
\code{\link{ImputeMechanism.object}}
}

