% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dejaData.R
\name{MakeDejaData}
\alias{MakeDejaData}
\title{Create a \code{DejaData} object}
\usage{
MakeDejaData(data, arm, rate, Id)
}
\arguments{
\item{data}{A data frame containing the subject}

\item{arm,}{character the column name of the treatment arm for each subject}

\item{rate,}{character the column name of the rate to be used when simulating}

\item{Id,}{character the column name of subject Id}
}
\description{
This object is can be used to create a \code{SingleSim} object with
subject specific rates
}

