% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit.R
\name{NBfit}
\alias{NBfit}
\title{S3 generic for fitting an negative binomial model}
\usage{
NBfit(x, equal.dispersion, formula, ...)
}
\arguments{
\item{x}{The S3 object}

\item{formula}{The formula to be used when calling \code{glm.nb} this should be
left as the default value for all but advanced users}

\item{...}{Additional arguments to be passed to the Surv function}
}
\description{
S3 generic for fitting an negative binomial model
}

