% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_engine.R
\name{Similarity}
\alias{Similarity}
\title{Similarity}
\usage{
Similarity(
  y,
  fh,
  LoadData,
  path,
  nts = 100,
  Preprocessing = 1,
  Dist_type = 3,
  gamma = 0.05,
  scaling = c("sdiff", "snaive"),
  clustering = 0,
  FullOutput = FALSE
)
}
\arguments{
\item{y}{"vector" the observation series}

\item{fh}{number of forecasting horizon}

\item{LoadData}{}

\item{nts}{number of the most similar series}

\item{Preprocessing}{remove seasonality, smooth the series and scale the target and the reference
series to the same magnitude}

\item{Dist_type}{number of forecasting benchmarks}

\item{gamma}{calibrated significance level quantiles}

\item{scaling}{scaling method: using different scaling or not}

\item{clustering}{number of sample categories in kmeans}

\item{FullOutput}{}
}
\value{
A list containing calculation of similarity
}
\description{
Similarity
}
