% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SingleSim.R
\name{SimulateDropout}
\alias{SimulateDropout}
\title{Simulate subject dropout}
\usage{
SimulateDropout(simComplete, drop.mechanism)
}
\arguments{
\item{simComplete}{TODO}

\item{drop.mechanism}{TODO}
}
\value{
A \code{SingleSim} object with \code{status='dropout'}
}
\description{
STUFF
}

