% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SingleSim.R
\name{SingleSim.object}
\alias{SingleSim.object}
\alias{print.SingleSim}
\alias{summary.SingleSim}
\title{SingleSim Object}
\arguments{
\item{data}{TODO}

\item{event.times}{TODO}

\item{status}{TODO}

\item{subject.rates}{TODO}

\item{dropout.mechanism}{TODO}

\item{study.time}{TODO}

\item{event.rates}{TODO}

\item{dispersions}{TODO}
}
\description{
Stuff
}
\details{
\code{print.SingleSim} and \code{summary.SingleSim} methods are defined.
}
\section{Structure}{
 The following components must be included in
a SingleSim Object
}

