% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SingleSimFit.R
\name{SingleSimFit.object}
\alias{SingleSimFit.object}
\title{SingleSimFit object}
\arguments{
\item{singleSim}{The \code{SingleSim} object to which a model has been fitted}

\item{model}{The model which has been fitted}

\item{genCoeff.function}{A function which returns a list of parameters from the model fit(s) which can
be used when performing the gamma imputation (these could be adapted to include uncertainty in these parameters)
If a Poisson/quasi-Poission model was fitted to the \code{SingleSimFit} object
then this will be NULL}

\item{equal}{dispersion whether equal dispersions were used when fitting model(s) to the data}
}
\description{
A \code{SingleSimFit} object is returned from calling \code{Simfit} with
a \code{SingleSim object}. It can be used to both impute data sets or can be summarized
}
\details{
A \code{\link{summary.SingleSimFit}} method has been implemented
}

