% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeSimFit.R
\name{summary.ImputeSimFit.object}
\alias{summary.ImputeSimFit.object}
\title{summary.ImputeSimFit object}
\arguments{
\item{treatment.effect}{The mean of the estimated treatment.effect
from the imputed data}

\item{se}{The standard error of the (log) treatment effect calculated using Rubin's formula}

\item{df}{The number of degrees of freedom used to calculate the p-value}

\item{adjusted.df}{The number of degrees of freedom used to calculate the adjusted
p-value (this should be used if the complete data number of degrees of freedom is small)}

\item{dispersion}{The mean of the estimated dispersion parameter}

\item{pval}{The p-value for the test log(treatment.effect)=0 using Rubin's formula}

\item{adjusted.pval}{The p-value for the test log(treatment.effect)=0 using Rubin's
formula and the adjusted number of degrees of freedom}

\item{dropout}{The number of subjects who drop out (per arm) for this imputed data set}

\item{number.subjects}{The number of subjects (per arm) for this imputed data set}
}
\description{
The summary of a \code{ImputeSimFit} object. Rubin's formula 
is used to combine the test statistics into a single summary
}
\details{
A \code{print.summary.ImputeSimFit} object has been implemented
}
