% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSim.R
\name{summary.SingleSim.object}
\alias{summary.SingleSim.object}
\title{summary.SingleSim object}
\arguments{
\item{status}{The status of the SingleSim object}

\item{study.time}{The study.time from the SingleSim object}

\item{number.subjects}{The number of subjects on each arm}

\item{number.dropouts}{The number of subjects who dropout on each arm}

\item{total.events}{The total number of events for each arm}

\item{time.at.risk}{The total time at risk for each arm}

\item{empirical.rates}{total.events/time.at.risk 

The \code{print.summary.SingleSim} method has been implemented}
}
\description{
The object returned when calling the summary function on a \code{SingleSim} object
}
