% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SingleSimFit.R
\name{summary.SingleSimFit}
\alias{summary.SingleSimFit}
\title{summary.SingleSimFit}
\arguments{
\item{model.summary}{The model summary from the fit}

\item{treatment.effect}{The estimate of treatment effect from the model fit}

\item{CI.limit}{The confidence interval limit (by default 0.95), call \code{summary(object,CI.limit=x)} to use
CI of \code{x} instead.}

\item{CI}{The confidence interval of the treatment effect}

\item{se}{Estimate for the standard error of (log) treatment effect}

\item{dispersion}{Estimate for the dispersion parameter or numeric(0) if Poisson/quasi-Poisson model used}

\item{rate.estimate}{Estimate of the event rates from the model a vector c(control arm, treatment arm)}

\item{pval}{The p value directly from the model fit (this is for the single model fit only, i.e. not using Rubin's formula)}

\item{datastatus}{The status of SingleSim object to which the fit was applied}
}
\description{
The summary object for a \code{SingleSimFit} object
}
\details{
A \code{print.summary.SingleSimFit} method has been implemented
}
\seealso{
\code{\link{SingleSimFit.object}}
}

