% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internalFit.R
\name{validateImputeArguments}
\alias{validateImputeArguments}
\title{check the arguments to Impute function are valid}
\usage{
validateImputeArguments(fit, impute.mechanism, N)
}
\description{
check the arguments to Impute function are valid
}

