% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{CreateScenario}
\alias{CreateScenario}
\title{Create \code{Scenario} object from list of Fit Summaries}
\usage{
CreateScenario(object, description = "")
}
\arguments{
\item{object}{Either a list of \code{summary.SingleSimFit}
or \code{summary.ImputeSimFit} objects}

\item{description}{A character string describing the scenario 
(used for printing)}
}
\value{
A \code{Scenario} object
}
\description{
Create \code{Scenario} object from list of Fit Summaries
}
\seealso{
\code{\link{Scenario.object}}
}
