% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeSim.R
\name{GetImputedDataSet}
\alias{GetImputedDataSet}
\title{Output a single imputed data set}
\usage{
GetImputedDataSet(imputeSim, index)
}
\arguments{
\item{imputeSim}{A \code{ImputeSim} object which contains 
multiple imputed data sets}

\item{index}{numeric, which of the multiple imputed data sets to output}
}
\value{
A \code{SingleSim} object with \code{status="imputed"}
}
\description{
Output a single imputed data set
}
\seealso{
\code{\link{ImputeSim.object}}
}
