% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{numberSubjects}
\alias{numberSubjects}
\title{S3 generic to output the number of subjects in a given object}
\usage{
numberSubjects(x)
}
\arguments{
\item{x}{The object}
}
\value{
The number of subjects
}
\description{
S3 generic to output the number of subjects in a given object
}
\examples{
sim <- SimulateComplete(study.time=365,number.subjects=50,
event.rates=c(0.01,0.005),dispersions=0.25)
numberSubjects(sim)
}
