% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common.R
\name{Simfit}
\alias{Simfit}
\title{S3 generic for fitting an negative binomial model}
\usage{
Simfit(x, equal.dispersion, formula, ...)
}
\arguments{
\item{x}{The S3 object}

\item{equal.dispersion}{logical, should the arms have the same dispersion parameter when
fitting the negative binomial model}

\item{formula}{The formula to be used when calling \code{glm.nb} this should be
left as the default value for all but advanced users}

\item{...}{Additional arguments to be passed to \code{glm} or \code{glm.nb}}
}
\value{
A \code{SingleSimFit} object
}
\description{
S3 generic for fitting an negative binomial model
}
\seealso{
\code{\link{SingleSimFit.object}}
}

