% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{subjectsPerArm}
\alias{subjectsPerArm}
\title{S3 generic to output the number of subjects in each arm for a given object}
\usage{
subjectsPerArm(x)
}
\arguments{
\item{x}{The object}
}
\value{
A vector of the number of subjects in each arm
}
\description{
S3 generic to output the number of subjects in each arm for a given object
}

